Refactor email channel: split IDLE supervisor and helpers
Extract the IDLE supervisor and formatting helpers out of email.ts so
each file stays under the 300-line guideline enforced on upstream.

- email-idle-loop.ts (294 lines): ImapIdleSupervisor class owning the
  client lifecycle (connect, IDLE-listen, 20-min refresh, reconnect
  with exponential backoff) plus the ImapFlowClient/ImapMessage types
- email-helpers.ts (77 lines): textToHtml, extractBodyText, isAutoReply
  (stateless pure functions, no IMAP/SMTP deps)
- email.ts (246 lines, was 497): EmailChannel class only - the Channel
  contract (send/receive/connect/disconnect)

Behavior is identical; the boundary moved. No public API change.

Also fixes a state-flip ordering issue uncovered during the split:
ImapIdleSupervisor.connect() now only builds the client, and a new
startLoop() method begins supervision. The owner must flip its state
to connected between the two calls so the supervisor's isConnected
hook returns true on the first loop check. Previously the inline
connect() started the loop before the state flip, which would have
caused the loop to exit immediately if startLoop had been async enough
to hit the first check before the synchronous state assignment.

All 13 existing tests pass; no test changes needed.
